+++ schema/User.js
@@ -1,0 +1,25 @@
+import { Schema, model } from "mongoose";
+import crypto from "crypto";
+
+const userSchema = new Schema({
+  name: {
+    type: String,
+    unique: true,
+    required: true,
+  },
+  email: {
+    type: String,
+    unique: true,
+    required: true,
+  },
+  password: {
+    type: String,
+    required: true
+  },
+  accessToken: {
+    type: String,
+    default: () => crypto.randomBytes(128).toString("hex")
+  }
+});
+
+export const User = model("User", userSchema);
